Move logic to separate class
