+++ docs/src/test/java/io/improbable/snippet/SequenceExample.java
@@ -152,7 +152,8 @@ public class SequenceExample {
 
         // Now you can actually use Keanu to reconstruct the Sequence object
         ProtobufLoader loader = new ProtobufLoader();
-        Sequence reconstructedSequence = loader.loadSequence(new FileInputStream(new File("file_name.proto")));
+        BayesianNetwork reconstructedNetwork = loader.loadNetwork(new FileInputStream(new File("file_name.proto")));
+        Sequence reconstructedSequence = Sequence.loadFromBayesNet(reconstructedNetwork);
         x1Retrieved = reconstructedSequence.asList().get(0).get(x1Label);
 
         //%%SNIPPET_END%% SequenceTimeSeries
